% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sf.R
\name{check_geopi_sf}
\alias{check_geopi_sf}
\title{Check if sf exists}
\usage{
check_geopi_sf(gdot_pi, pi_check = TRUE)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can be a single PI or a vector or list of PIs. If more than one PI is provided, each project will get its own row.}

\item{pi_check}{Check if the PI is a valid format. Defaults to TRUE.}
}
\value{
A tibble
}
\description{
Check if shapefile available on GeoPI without downloading
}
\examples{
check_geopi_sf(c("0015478","0001111"))
}
