% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pi.R
\name{check_pi}
\alias{check_pi}
\title{Check if PI is possible based on format}
\usage{
check_pi(gdot_pi)
}
\arguments{
\item{gdot_pi}{GDOT PI. If not seven digits, will be padded if it's numeric}
}
\value{
True/False
}
\description{
GDOT PIs are seven character IDs formatted as 1) numbers with leading zeros, 2) six numbers and a "-", and 3) six numbers with a leading M or S. Checks the format to skip running a GeoPI call if the project ID is not possible for the system.
}
\examples{
check_pi("0012345")
check_pi("546540-")
check_pi("M023424")
check_pi("FAKE523")
}
