% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopi_extract_sf.R
\name{cr_check}
\alias{cr_check}
\title{Check for Concept Reports}
\usage{
cr_check(gdot_pi)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can take a list or vector of IDs.}
}
\value{
tibble
}
\description{
Check for Concept Reports
}
\examples{
\dontrun{
cr_check(gdot_pi = "0000820")
}
}
