% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopi_session.R
\name{geopi_session}
\alias{geopi_session}
\title{geopi session}
\usage{
geopi_session(...)
}
\arguments{
\item{...}{any arguments passed to `polite::bow` can be passed to}
}
\value{
object of class `polite`, `session`
}
\description{
`geopi_session` masks `polite::bow`, creating a polite session with GeoPI. This can be plugged into any of the `get_geopi` family of functions.
}
\examples{
\donttest{
session <- geopi_session()
session
}
}
