% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopi_extract_sf.R
\name{get_geopi}
\alias{get_geopi}
\title{Get GeoPI Data}
\usage{
get_geopi(gdot_pi, geometry = FALSE)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can be a single PI or a vector or list of PIs. If more than one PI is provided, each project will get its own row.}

\item{geometry}{if FALSE (the default), return a regular tibble of GeoPI data. if TRUE, uses the get_geopi_sf function to return an sf tibble with simple feature geometry in the 'geometry' column.}
}
\value{
A tibble or an sf tibble depending on if geometry is FALSE or TRUE
}
\description{
Get the available GeoPI data for a project. Adds the field `Gather.Date` with the current date to document the date of extraction.
}
