% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geopi_ef.R
\name{get_geopi_docs_ef}
\alias{get_geopi_docs_ef}
\title{Check for Concept Reports}
\usage{
get_geopi_docs_ef(
  gdot_pi,
  page_scrape,
  mode = c("cr_only", "cr_check", "doc_summary"),
  gather_date = NULL
)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can take a list or vector of IDs.}

\item{page_scrape}{Scraped GeoPI Page supplied by `get_geopi_ef`.}

\item{mode}{if "cr_only" (the default), returns the filepath and filenames of all approved concept reports. If "cr_check", merely returns T/F for each project if it has a concept report. If "doc_summary", returns the file name, type, and path for all project documents.}

\item{gather_date}{Date information is gathered from GeoPI. Defaults to today.}
}
\value{
a tibble
}
\description{
Check for Concept Reports
}
\examples{
\dontrun{
get_geopi_docs_ef(gdot_pi = "0000820", mode = "cr_check")
}
}
