% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geopi_ef.R
\name{get_geopi_overview_ef}
\alias{get_geopi_overview_ef}
\title{get_geopi_overview}
\usage{
get_geopi_overview_ef(gdot_pi, page_scrape, gather_date = NULL)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can take a list or vector of IDs.}

\item{page_scrape}{Scraped GeoPI Page supplied by `get_geopi_ef`.}

\item{gather_date}{Date information is gathered from GeoPI. Defaults to today.}
}
\value{
a tibble
}
\description{
get_geopi_overview
}
\examples{
\dontrun{
get_geopi_overview_ef(gdot_pi = "0000820")
}
}
