% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopi_extract_sf.R
\name{get_geopi_phase}
\alias{get_geopi_phase}
\title{Get Project Phase}
\usage{
get_geopi_phase(gdot_pi, session = NULL)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can take a list or vector of IDs.}

\item{session}{If NULL (the default), creates a new session. Can provide a session made with `polite::bow`.}
}
\value{
a tibble
}
\description{
Get Project Phase
}
\examples{
\dontrun{
get_geopi_phase(gdot_pi = "0000820")
}
}
