% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopi_extract_sf.R
\name{get_geopi_sf}
\alias{get_geopi_sf}
\title{Get GeoPI Spatial Data}
\usage{
get_geopi_sf(gdot_pi)
}
\arguments{
\item{gdot_pi}{GDOT Project ID}
}
\value{
An sf tibble
}
\description{
Get the spatial geometry of a GDOT project. `get_geopi_sf2` uses the httr2 package.
}
