% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeopi-package.R
\docType{package}
\name{rgeopi-package}
\alias{rgeopi}
\alias{rgeopi-package}
\title{rgeopi: Gather Data from GDOT's GeoPI Database}
\description{
rGeoPI allows the user to easily interface with GDOT's GeoPI repository. This includes project information, phase years and costs, project geometry, and document information, returned in data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ModernMobilityPartners/rgeopi}
  \item Report bugs at \url{https://github.com/ModernMobilityPartners/rgeopi/issues}
}

}
\author{
\strong{Maintainer}: Freyja Brandel-Tanis \email{fbrandel-tanis@modernmobilitypartners.com} (\href{https://orcid.org/0000-0003-2870-6980}{ORCID})

Other contributors:
\itemize{
  \item  Modern Mobility Partners, LLC [copyright holder]
}

}
\keyword{internal}
