% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place.R
\name{format_place_gis}
\alias{format_place_gis}
\title{Format sf for PLA+CE GIS Tool}
\usage{
format_place_gis(project_sf)
}
\arguments{
\item{project_sf}{sf object of project geometry with Project ID in column "Project.ID" - Result of `get_geopi_sf`}
}
\value{
sf object
}
\description{
Format sf for PLA+CE GIS Tool
}
