% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geopi.R
\name{get_geopi_docs}
\alias{get_geopi_docs}
\title{Check for Concept Reports}
\usage{
get_geopi_docs(
  gdot_pi,
  session = NULL,
  mode = c("cr_only", "cr_check", "doc_summary"),
  gather_date = NULL
)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can take a list or vector of IDs.}

\item{session}{If NULL (the default), creates a new session. Can provide a session made with `polite::bow`.}

\item{mode}{if "cr_only" (the default), returns the filepath and filenames of all approved concept reports. If "cr_check", merely returns T/F for each project if it has a concept report. If "doc_summary", returns the file name, type, and path for all project documents.}
}
\value{
a tibble
}
\description{
Check for Concept Reports
}
\examples{
\dontrun{
get_geopi_docs(gdot_pi = "0000820", mode = "cr_check")
}
}
