% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geopi.R
\name{get_geopi_sf}
\alias{get_geopi_sf}
\title{Get GeoPI Spatial Data}
\usage{
get_geopi_sf(gdot_pi)
}
\arguments{
\item{gdot_pi}{GDOT Project ID. Can be a single PI or a vector or list of PIs. If more than one PI is provided, each project will get its own row. get_geopi_sf2 can only take a single}
}
\value{
An sf tibble
}
\description{
Get the spatial geometry of a GDOT project. `get_geopi_sf2` uses the httr2 package.
}
\examples{
\dontrun{
get_geopi_sf(gdot_pi = "0000820")
}
}
